+++ yolox/evaluators/mot_evaluator.py
@@ -104,7 +104,9 @@ class MOTEvaluator:
         tensor_type = torch.cuda.HalfTensor if half else torch.cuda.FloatTensor
         model = model.eval()
         if half:
-            model = model.half()
+            # model = model.half()
+            model = model.float()
+
         ids = []
         data_list = []
         results = []
@@ -248,7 +250,8 @@ class MOTEvaluator:
         tensor_type = torch.cuda.HalfTensor if half else torch.cuda.FloatTensor
         model = model.eval()
         if half:
-            model = model.half()
+            # model = model.half()
+            model = model.float()
         ids = []
         data_list = []
         results = []
@@ -371,7 +374,8 @@ class MOTEvaluator:
         tensor_type = torch.cuda.HalfTensor if half else torch.cuda.FloatTensor
         model = model.eval()
         if half:
-            model = model.half()
+            # model = model.half()
+            model = model.float()
         ids = []
         data_list = []
         results = []
@@ -494,7 +498,8 @@ class MOTEvaluator:
         tensor_type = torch.cuda.HalfTensor if half else torch.cuda.FloatTensor
         model = model.eval()
         if half:
-            model = model.half()
+            # model = model.half()
+            model = model.float()
         ids = []
         data_list = []
         results = []
